refactor: move WalletName to Utils dir
