+++ app/components/Home/index.js
@@ -7,4 +7,3 @@ export WalletUnlocker from './WalletUnlocker'
 export WalletSettingsFormLocal from './WalletSettingsFormLocal'
 export WalletSettingsFormRemote from './WalletSettingsFormRemote'
 export WalletHeader from './WalletHeader'
-export WalletName from './WalletName'
