+++ app/components/Util/WalletName.js
