+++ app/components/Util/index.js
@@ -1,0 +1,1 @@
+export WalletName from './WalletName'
